Bring the record types back from the records directory to entity-types directory
