+++ packages/core-data/src/entity-types/attachment.ts
