+++ packages/core-data/src/entity-types/menu-location.ts
