+++ packages/core-data/src/entity-types/nav-menu-item.ts
