+++ packages/core-data/src/entity-types/navigation-area.ts
