+++ packages/core-data/src/entity-types/plugin.ts
