+++ packages/core-data/src/entity-types/post.ts
