+++ packages/core-data/src/entity-types/settings.ts
