+++ packages/core-data/src/entity-types/sidebar.ts
