+++ packages/core-data/src/entity-types/taxonomy.ts
