+++ packages/core-data/src/entity-types/widget-type.ts
