+++ packages/core-data/src/entity-types/widget.ts
