+++ packages/core-data/src/entity-types/wp-template-part.ts
